Эй, если вдруг меня не станет, брат (брат)
a

+++ Assets/Scripts/LightMovement.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class LightMovement : MonoBehaviour
 {
     public new Transform transform;
-    public int rotationAngle;
+    public float rotationAngle;
     public void FixedUpdate()
     {
         transform.Rotate(new Vector3(rotationAngle, 0, 0) * Time.deltaTime);
